refactor: App.js 생성 및 index.js 요소 이동
