+++ src/domain/Lotto.js
@@ -6,11 +6,14 @@ class Lotto {
   getCorretNumber(targetNumber) {
     let correctNumber = 0;
     for (let i = 0; i < 6; i += 1) {
-      if (targetNumber.includes(this.numbers[i])) {
-        correctNumber += 1;
-      }
+      correctNumber += this.#addCorrectNumber(targetNumber, this.numbers[i]);
     }
     return correctNumber;
+  }
+
+  #addCorrectNumber(targetNumber, number) {
+    if (targetNumber.includes(number)) return 1;
+    return 0;
   }
 
   hasBonusNumber(bonusNumber) {
